refactor(report): move the tab sections into a templates tier
