+++ packages/nestjs-doctor/src/report/ui/templates/tab-diagnosis.ts
