+++ packages/nestjs-doctor/src/report/ui/templates/tab-endpoints.ts
