+++ packages/nestjs-doctor/src/report/ui/templates/tab-lab.ts
