+++ packages/nestjs-doctor/src/report/ui/templates/tab-modules-graph.ts
