+++ packages/nestjs-doctor/src/report/ui/templates/tab-schema.ts
