+++ packages/nestjs-doctor/src/report/ui/templates/tab-summary.ts
